refactor(racingCar): package 이름 변경
- entity에서 vo로 변경
