+++ src/main/java/racingcar/model/domain/Racing.java
@@ -1,8 +1,7 @@
 package racingcar.model.domain;
 
 import java.util.List;
-import java.util.stream.Collectors;
-import racingcar.model.entity.Car;
+import racingcar.model.vo.Car;
 
 public class Racing {
